% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday_lst.R
\name{holiday_lst}
\alias{holiday_lst}
\title{Fedral Holidays DataFrame}
\usage{
holiday_lst()
}
\value{
Returns a data frame if fedral holidays and when they were observed from 2012 - 2020
}
\description{
This is a simple function that returns a data frame of federal holidays and date on which they were observed. It has only 2 columns, Date and Holiday for years 2012-2020.The functions does not take any parameter but returns a data frame.
}
\examples{
holiday_lst()
}
