% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_gender.R
\name{infer_gender}
\alias{infer_gender}
\title{Infer gender for list of names fast}
\usage{
infer_gender(x, detect_cores)
}
\arguments{
\item{x}{The vector of names to perform paralleized inference on}

\item{detect_cores}{If False (default), 2 cores are used. If True, half the number of cores are used on Mac OS else 2 cores are used.}
}
\value{
The a vector with m/f/u if the name in the original vector was male female or Unknown respectively
}
\description{
This function used a database of names and corresponding gender to lookup the gender of a name. It takes in a vector of names and returns a vector indicating gender for the name. The gender is denoted by "m"/"f"/"u" for male/female/unknown respectively.
}
\examples{
infer_gender(x=c("abby", "Nick", "abc"))
}
